+++ shared/package.json
@@ -15,5 +15,10 @@
     "link": "echo no operation",
     "unlink": "echo no operation",
     "publish-package": "echo no operation"
+  },
+  "devDependencies": {
+    "@types/node": "^22.13.10",
+    "tsx": "^4.19.3",
+    "typescript": "^5.8.2"
   }
 }
